+++ tests/unit/test_charm.py
@@ -26,47 +26,13 @@ def test_config_changed_no_haproxy_route_relation(context):
     assert out.unit_status == ops.testing.BlockedStatus("Missing haproxy-route relation.")
 
 
-def test_config_changed_invalid_mode(monkeypatch: pytest.MonkeyPatch, context):
-    """
-    arrange: prepare some state and patch invalid mode.
-    act: trigger a config changed event.
-    assert: status is blocked.
-    """
-    monkeypatch.setattr(state, "get_mode", MagicMock(side_effect=state.UndefinedModeError))
-    charm_state = ops.testing.State(
-        relations=[ops.testing.Relation("haproxy-route")],
-    )
-
-    out = context.run(context.on.config_changed(), charm_state)
-
-    assert out.unit_status == ops.testing.BlockedStatus("Operating mode is undefined.")
-
-
-def test_config_changed_integrator_missing_config(monkeypatch: pytest.MonkeyPatch, context):
-    """
-    arrange: prepare some state with haproxy-route relation and no configuration.
-    act: trigger a config changed event.
-    assert: status is blocked.
-    """
-    monkeypatch.setattr(state, "get_mode", MagicMock(return_value=state.Mode.INTEGRATOR))
-    charm_state = ops.testing.State(
-        relations=[ops.testing.Relation("haproxy-route")],
-    )
-
-    out = context.run(context.on.config_changed(), charm_state)
-
-    assert out.unit_status == ops.testing.BlockedStatus(
-        "Missing configuration for integrator mode: backend-addresses backend-ports"
-    )
-
-
-def test_config_changed_integrator_invalid_addresses(monkeypatch: pytest.MonkeyPatch, context):
+def test_config_changed_invalid_state(monkeypatch: pytest.MonkeyPatch, context):
     """
     arrange: prepare some state with invalid backend-addresses.
     act: trigger a config changed event.
-    assert: status is bllocked.
+    assert: status is blocked.
     """
-    monkeypatch.setattr(state, "get_mode", MagicMock(return_value=state.Mode.INTEGRATOR))
+    monkeypatch.setattr(state.State, "from_charm", MagicMock(side_effect=state.InvalidStateError))
     charm_state = ops.testing.State(
         config={"backend-addresses": "10.0.0.1,invalid", "backend-ports": "8080,8081"},
         relations=[ops.testing.Relation("haproxy-route")],
@@ -74,25 +40,20 @@ def test_config_changed_integrator_invalid_addresses(monkeypatch: pytest.MonkeyP
 
     out = context.run(context.on.config_changed(), charm_state)
 
-    assert out.unit_status == ops.testing.BlockedStatus(
-        "Invalid integrator configuration: backend_addresses-1"
-    )
+    assert out.unit_status == ops.testing.BlockedStatus()
 
 
-def test_config_changed_integrator_invalid_ports(monkeypatch: pytest.MonkeyPatch, context):
+def test_config_changed_integrator(context):
     """
-    arrange: prepare some state with invalid ports.
-    act: trigger a config changed event with invalid port configuration.
-    assert: status is blocked with the correct message.
+    arrange: prepare some valid state for an integrator.
+    act: trigger a config changed event.
+    assert: status is active.
     """
-    monkeypatch.setattr(state, "get_mode", MagicMock(return_value=state.Mode.INTEGRATOR))
     charm_state = ops.testing.State(
-        config={"backend-addresses": "10.0.0.1,10.0.0.2", "backend-ports": "99999"},
+        config={"backend-addresses": "10.0.0.1,10.0.0.2", "backend-ports": "8080,8081"},
         relations=[ops.testing.Relation("haproxy-route")],
     )
 
     out = context.run(context.on.config_changed(), charm_state)
 
-    assert out.unit_status == ops.testing.BlockedStatus(
-        "Invalid integrator configuration: backend_ports-0"
-    )
+    assert out.unit_status == ops.testing.ActiveStatus()