+++ tests/unit/test_state.py
@@ -6,110 +6,104 @@
 from unittest.mock import Mock
 
 import pytest
-from ops import CharmBase, Relation
+from charms.traefik_k8s.v2.ingress import (
+    IngressRequirerAppData,
+    IngressRequirerData,
+    IngressRequirerUnitData,
+)
+from ops import CharmBase
 
 import state
 
 
-def test_get_mode_integrator():
+def test_adapter_state_from_charm():
     """
-    arrange: mock a charm with backend configuration
-    act: get mode
-    assert: mode is 'integrator'
-    """
-    charm = Mock(CharmBase)
-    charm.config = {
-        "backend-addresses": "127.0.0.1,127.0.0.2",
-        "backend-ports": "8080,8081",
-    }
-    assert state.Mode.INTEGRATOR == state.get_mode(charm, None)
-
-
-def test_get_mode_adapter():
-    """
-    arrange: mock a charm without backend configuration and a relation
-    act: get mode
-    assert: mode is 'adapter'
-    """
-    charm = Mock(CharmBase)
-    charm.config = {}
-    relation = Mock(Relation)
-    assert state.Mode.ADAPTER == state.get_mode(charm, relation)
-
-
-def test_get_mode_invalid():
-    """
-    arrange: mock a charm with backend configuration and a relation
-    act: get mode
-    assert: a UndefinedModeError is raised
+    arrange: mock a charm with an ingress relation
+    act: instantiate a State
+    assert: the data matches the charm configuration
     """
     charm = Mock(CharmBase)
     charm.config = {
-        "backend-addresses": "127.0.0.1,127.0.0.2",
-        "backend-ports": "8080,8081",
+        "retry-count": 1,
+        "retry-interval": 10,
+        "retry-redispatch": True,
+        "paths": "/api/v1,/api/v2",
+        "subdomains": "api",
     }
-    relation = Mock(Relation)
-    with pytest.raises(state.UndefinedModeError):
-        state.get_mode(charm, relation)
+    ingress_relation_data = IngressRequirerData(
+        app=IngressRequirerAppData(model="model", name="name", port=8080),
+        units=[IngressRequirerUnitData(host="sample.host", ip="127.0.0.1")],
+    )
+    charm_state = state.State.from_charm(charm, ingress_relation_data)
+    assert [str(address) for address in charm_state.backend_addresses] == [
+        ingress_relation_data.units[0].ip
+    ]
+    assert charm_state.backend_ports == [ingress_relation_data.app.port]
+    assert charm_state.retry_count == charm.config.get("retry-count")
+    assert charm_state.retry_interval == charm.config.get("retry-interval")
+    assert charm_state.retry_redispatch == charm.config.get("retry-redispatch")
+    assert charm_state.paths == charm.config.get("paths").split(",")
+    assert charm_state.subdomains == charm.config.get("subdomains").split(",")
 
 
-def test_get_integrator_information():
+def test_integrator_state_from_charm():
     """
     arrange: mock a charm with backend configuration
-    act: instantiate a IntegratorInformation
+    act: instantiate a State
     assert: the data matches the charm configuration
     """
     charm = Mock(CharmBase)
     charm.config = {
         "backend-addresses": "127.0.0.1,127.0.0.2",
         "backend-ports": "8080,8081",
-        "paths": "/api/v1,/api/v2",
-        "subdomains": "api",
+        "retry-count": 1,
+        "retry-interval": 10,
+        "retry-redispatch": True,
     }
-    info = state.IntegratorInformation.from_charm(charm)
-    assert [str(address) for address in info.backend_addresses] == charm.config.get(
+    charm_state = state.State.from_charm(charm, None)
+    assert [str(address) for address in charm_state.backend_addresses] == charm.config.get(
         "backend-addresses"
     ).split(",")
-    assert [str(port) for port in info.backend_ports] == charm.config.get("backend-ports").split(
-        ","
-    )
-    assert info.paths == charm.config.get("paths", "").split(",")
-    assert info.subdomains == charm.config.get("subdomains", "").split(",")
+    assert [str(port) for port in charm_state.backend_ports] == charm.config.get(
+        "backend-ports"
+    ).split(",")
+    assert charm_state.retry_count == charm.config.get("retry-count")
+    assert charm_state.retry_interval == charm.config.get("retry-interval")
+    assert charm_state.retry_redispatch == charm.config.get("retry-redispatch")
 
 
-def test_get_integrator_information_no_address():
+def test_state_from_charm_no_backend():
+    """
+    arrange: mock a charm with backend address and without backend configuration not ingress
+    act: instantiate a State
+    assert: a InvalidStateError is raised
+    """
+    charm = Mock(CharmBase)
+    charm.config = {}
+    with pytest.raises(state.InvalidStateError):
+        state.State.from_charm(charm, None)
+
+
+def test_state_from_charm_invalid_address():
     """
     arrange: mock a charm with backend port and without address configuration
-    act: instantiate a IntegratorInformation
-    assert: a InvalidIntegratorConfigError is raised
-    """
-    charm = Mock(CharmBase)
-    charm.config = {
-        "backend-ports": "8080,8081",
-    }
-    with pytest.raises(state.InvalidIntegratorConfigError):
-        state.IntegratorInformation.from_charm(charm)
-
-
-def test_get_integrator_information_no_port():
-    """
-    arrange: mock a charm with backend address and without port configuration
-    act: instantiate a IntegratorInformation
-    assert: a InvalidIntegratorConfigError is raised
-    """
-    charm = Mock(CharmBase)
-    charm.config = {
-        "backend-addresses": "127.0.0.1,127.0.0.2",
-    }
-    with pytest.raises(state.InvalidIntegratorConfigError):
-        state.IntegratorInformation.from_charm(charm)
-
-
-def test_config_changed_integrator_invalid_paths():
-    """
-    arrange: mock a charm with backend address, port configuration and invalid paths.
     act: instantiate a State
-    assert: a InvalidIntegratorConfigError is raised
+    assert: a InvalidStateError is raised
+    """
+    charm = Mock(CharmBase)
+    charm.config = {
+        "backend-addresses": "invalid",
+        "backend-ports": "8080",
+    }
+    with pytest.raises(state.InvalidStateError):
+        state.State.from_charm(charm, None)
+
+
+def test_state_from_charm_invalid_paths():
+    """
+    arrange: mock a charm with backend addresses, ports configuration and invalid paths
+    act: instantiate a State
+    assert: a InvalidStateError is raised
     """
     charm = Mock(CharmBase)
     charm.config = {
@@ -117,15 +111,58 @@ def test_config_changed_integrator_invalid_paths():
         "backend-ports": "8080,8081",
         "paths": "invalid path",
     }
-    with pytest.raises(state.InvalidIntegratorConfigError):
-        state.IntegratorInformation.from_charm(charm)
+    with pytest.raises(state.InvalidStateError):
+        state.State.from_charm(charm, None)
 
 
-def test_config_changed_integrator_invalid_subdomains():
+def test_state_from_charm_invalid_port():
     """
-    arrange: mock a charm with backend address, port configuration and invalid subdomains.
+    arrange: mock a charm with backend address and without port configuration
     act: instantiate a State
-    assert: a InvalidIntegratorConfigError is raised
+    assert: a InvalidStateError is raised
+    """
+    charm = Mock(CharmBase)
+    charm.config = {
+        "backend-addresses": "127.0.0.1,127.0.0.2",
+        "backend-ports": "99999",
+    }
+    with pytest.raises(state.InvalidStateError):
+        state.State.from_charm(charm, None)
+
+
+def test_state_from_charm_invalid_retry_count():
+    """
+    arrange: mock a charm with backend address and without retry-count configuration
+    act: instantiate a State
+    assert: a InvalidStateError is raised
+    """
+    charm = Mock(CharmBase)
+    charm.config = {
+        "retry-count": -1,
+    }
+    with pytest.raises(state.InvalidStateError):
+        state.State.from_charm(charm, None)
+
+
+def test_state_from_charm_invalid_retry_interval():
+    """
+    arrange: mock a charm with backend address and without retry-interval configuration
+    act: instantiate a State
+    assert: a InvalidStateError is raised
+    """
+    charm = Mock(CharmBase)
+    charm.config = {
+        "retry-interval": -1,
+    }
+    with pytest.raises(state.InvalidStateError):
+        state.State.from_charm(charm, None)
+
+
+def test_state_from_charm_invalid_subdomains():
+    """
+    arrange: mock a charm with backend addresses, ports configuration and invalid subdomains
+    act: instantiate a State
+    assert: a InvalidStateError is raised
     """
     charm = Mock(CharmBase)
     charm.config = {
@@ -133,5 +170,5 @@ def test_config_changed_integrator_invalid_subdomains():
         "backend-ports": "8080,8081",
         "subdomains": "invalid$subdomains",
     }
-    with pytest.raises(state.InvalidIntegratorConfigError):
-        state.IntegratorInformation.from_charm(charm)
+    with pytest.raises(state.InvalidStateError):
+        state.State.from_charm(charm, None)